+++ ambry-mysql/src/main/java/com/github/ambry/accountstats/ContainerUsageFunction.java
@@ -11,9 +11,20 @@
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  */
-package com.github.ambry.server.mysql;
+package com.github.ambry.accountstats;
 
+/**
+ * A callback function to call when processing container storage usage.
+ */
 @FunctionalInterface
-public interface AggregatedContainerUsageFunction {
-  void apply(short accountId, short containerId, long storageUsage);
+public interface ContainerUsageFunction {
+
+  /**
+   * Process container storage usage.
+   * @param partitionId
+   * @param accountId
+   * @param containerId
+   * @param storageUsage
+   */
+  void apply(short partitionId, short accountId, short containerId, long storageUsage, long updatedAtMs);
 }
