+++ ambry-mysql/src/main/resources/AmbryContainerStorageStats.ddl
