+++ ambry-server/src/main/java/com/github/ambry/server/AmbryServer.java
@@ -18,6 +18,8 @@ import com.codahale.metrics.jmx.JmxReporter;
 import com.github.ambry.account.AccountService;
 import com.github.ambry.account.AccountServiceCallback;
 import com.github.ambry.account.AccountServiceFactory;
+import com.github.ambry.accountstats.AccountStatsMySqlStore;
+import com.github.ambry.accountstats.AccountStatsMySqlStoreFactory;
 import com.github.ambry.clustermap.ClusterAgentsFactory;
 import com.github.ambry.clustermap.ClusterMap;
 import com.github.ambry.clustermap.ClusterParticipant;
@@ -69,8 +71,6 @@ import com.github.ambry.store.MessageInfo;
 import com.github.ambry.store.StorageManager;
 import com.github.ambry.store.StoreKeyConverterFactory;
 import com.github.ambry.store.StoreKeyFactory;
-import com.github.ambry.server.mysql.AccountStatsMySqlStore;
-import com.github.ambry.server.mysql.AccountStatsMySqlStoreFactory;
 import com.github.ambry.utils.SystemTime;
 import com.github.ambry.utils.Time;
 import com.github.ambry.utils.Utils;
